# Coucher.WebApi/swagger-live.json
{
  "openapi": "3.0.1",
  "info": {
    "title": "Coucher.WebApi",
    "version": "1.0"
  },
  "paths": {
...
    "/api/admin/task-templates": {
      "post": {
        "tags": [
          "AdminTaskTemplates"
        ],
        "parameters": [
          {
            "name": "AUGUSTUS_CORRELATION_ID",
            "in": "header"
          },
          {
            "name": "MAMRAZ_CORRELATION_ID",
            "in": "header"
          },
          {
            "name": "AUGUSTUS_CORRELATION_CREATION_TIMESTAMP",
            "in": "header"
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateTaskTemplateRequest"
              },
              "example": {
                "templateKey": "root-template",
                "seriesId": "3fa85f64-5717-4562-b3fc-2c963f66afa6",
                "categoryId": "3fa85f64-5717-4562-b3fc-2c963f66afb0",
                "name": "Root Template",
                "description": "Top level task template",
                "notes": "Optional notes",
                "defaultWeeksBeforeExerciseStart": 2,
                "influencerIds": [
                  "3fa85f64-5717-4562-b3fc-2c963f66afc1"
                ],
                "children": [
                  {
                    "templateKey": "child-a",
                    "name": "Child A",
                    "defaultWeeksBeforeExerciseStart": 1
                  },
                  {
                    "templateKey": "child-b",
                    "name": "Child B",
                    "defaultWeeksBeforeExerciseStart": 1,
                    "dependsOnTemplateKeys": [
                      "child-a"
                    ]
                  }
                ]
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateTaskTemplateRequest"
              },
              "example": {
                "templateKey": "root-template",
                "seriesId": "3fa85f64-5717-4562-b3fc-2c963f66afa6",
                "categoryId": "3fa85f64-5717-4562-b3fc-2c963f66afb0",
                "name": "Root Template",
                "description": "Top level task template",
                "notes": "Optional notes",
                "defaultWeeksBeforeExerciseStart": 2,
                "influencerIds": [
                  "3fa85f64-5717-4562-b3fc-2c963f66afc1"
                ],
                "children": [
                  {
                    "templateKey": "child-a",
                    "name": "Child A",
                    "defaultWeeksBeforeExerciseStart": 1
                  },
                  {
                    "templateKey": "child-b",
                    "name": "Child B",
                    "defaultWeeksBeforeExerciseStart": 1,
                    "dependsOnTemplateKeys": [
                      "child-a"
                    ]
                  }
                ]
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/CreateTaskTemplateRequest"
              },
              "example": {
                "templateKey": "root-template",
                "seriesId": "3fa85f64-5717-4562-b3fc-2c963f66afa6",
                "categoryId": "3fa85f64-5717-4562-b3fc-2c963f66afb0",
                "name": "Root Template",
                "description": "Top level task template",
                "notes": "Optional notes",
                "defaultWeeksBeforeExerciseStart": 2,
                "influencerIds": [
                  "3fa85f64-5717-4562-b3fc-2c963f66afc1"
                ],
                "children": [
                  {
                    "templateKey": "child-a",
                    "name": "Child A",
                    "defaultWeeksBeforeExerciseStart": 1
                  },
                  {
                    "templateKey": "child-b",
                    "name": "Child B",
                    "defaultWeeksBeforeExerciseStart": 1,
                    "dependsOnTemplateKeys": [
                      "child-a"
                    ]
                  }
                ]
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/TaskTemplate"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TaskTemplate"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/TaskTemplate"
                }
              }
            }
          }
        }
      }
    },
    "/api/admin/task-templates/bulk": {
      "post": {
        "tags": [
          "AdminTaskTemplates"
        ],
        "parameters": [
          {
            "name": "AUGUSTUS_CORRELATION_ID",
            "in": "header"
          },
          {
            "name": "MAMRAZ_CORRELATION_ID",
            "in": "header"
          },
          {
            "name": "AUGUSTUS_CORRELATION_CREATION_TIMESTAMP",
            "in": "header"
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "array",
                "items": {
                  "$ref": "#/components/schemas/CreateTaskTemplateRequest"
                }
              }
            },
            "text/json": {
              "schema": {
                "type": "array",
                "items": {
                  "$ref": "#/components/schemas/CreateTaskTemplateRequest"
                }
              }
            },
            "application/*+json": {
              "schema": {
                "type": "array",
                "items": {
                  "$ref": "#/components/schemas/CreateTaskTemplateRequest"
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/TaskTemplate"
                  }
                }
              },
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/TaskTemplate"
                  }
                }
              },
              "text/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/TaskTemplate"
                  }
                }
              }
            }
          }
        }
      }
    },
    "/api/admin/task-templates/{id}": {
      "put": {
        "tags": [
          "AdminTaskTemplates"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          },
          {
            "name": "AUGUSTUS_CORRELATION_ID",
            "in": "header"
          },
          {
            "name": "MAMRAZ_CORRELATION_ID",
            "in": "header"
          },
          {
            "name": "AUGUSTUS_CORRELATION_CREATION_TIMESTAMP",
            "in": "header"
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateTaskTemplateRequest"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateTaskTemplateRequest"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateTaskTemplateRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/TaskTemplate"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TaskTemplate"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/TaskTemplate"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/admin/task-templates/{id}/children": {
      "post": {
        "tags": [
          "AdminTaskTemplates"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          },
          {
            "name": "AUGUSTUS_CORRELATION_ID",
            "in": "header"
          },
          {
            "name": "MAMRAZ_CORRELATION_ID",
            "in": "header"
          },
          {
            "name": "AUGUSTUS_CORRELATION_CREATION_TIMESTAMP",
            "in": "header"
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateTaskTemplateChildRequest"
              },
              "example": {
                "templateKey": "child-b",
                "name": "Child B",
                "defaultWeeksBeforeExerciseStart": 1,
                "dependsOnTemplateKeys": [
                  "child-a"
                ]
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateTaskTemplateChildRequest"
              },
              "example": {
                "templateKey": "child-b",
                "name": "Child B",
                "defaultWeeksBeforeExerciseStart": 1,
                "dependsOnTemplateKeys": [
                  "child-a"
                ]
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/CreateTaskTemplateChildRequest"
              },
              "example": {
                "templateKey": "child-b",
                "name": "Child B",
                "defaultWeeksBeforeExerciseStart": 1,
                "dependsOnTemplateKeys": [
                  "child-a"
                ]
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/TaskTemplate"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TaskTemplate"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/TaskTemplate"
                }
              }
            }
          }
        }
      }
    },
    "/api/admin/task-templates/{id}/dependencies": {
      "post": {
        "tags": [
          "AdminTaskTemplates"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          },
          {
            "name": "AUGUSTUS_CORRELATION_ID",
            "in": "header"
          },
          {
            "name": "MAMRAZ_CORRELATION_ID",
            "in": "header"
          },
          {
            "name": "AUGUSTUS_CORRELATION_CREATION_TIMESTAMP",
            "in": "header"
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "string",
                "format": "uuid"
              }
            },
            "text/json": {
              "schema": {
                "type": "string",
                "format": "uuid"
              }
            },
            "application/*+json": {
              "schema": {
                "type": "string",
                "format": "uuid"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/TaskTemplateDependency"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TaskTemplateDependency"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/TaskTemplateDependency"
                }
              }
            }
          }
        }
      }
    },
    "/api/admin/task-templates/{id}/influencers": {
      "post": {
        "tags": [
          "AdminTaskTemplates"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          },
          {
            "name": "AUGUSTUS_CORRELATION_ID",
            "in": "header"
          },
          {
            "name": "MAMRAZ_CORRELATION_ID",
            "in": "header"
          },
          {
            "name": "AUGUSTUS_CORRELATION_CREATION_TIMESTAMP",
            "in": "header"
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "array",
                "items": {
                  "type": "string",
                  "format": "uuid"
                }
              }
            },
            "text/json": {
              "schema": {
                "type": "array",
                "items": {
                  "type": "string",
                  "format": "uuid"
                }
              }
            },
            "application/*+json": {
              "schema": {
                "type": "array",
                "items": {
                  "type": "string",
                  "format": "uuid"
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/TaskTemplateInfluencer"
                  }
                }
              },
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/TaskTemplateInfluencer"
                  }
                }
              },
              "text/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/TaskTemplateInfluencer"
                  }
                }
              }
            }
          }
        }
      }
    },
    "/api/admin/task-templates/dependencies/{dependencyId}": {
      "delete": {
        "tags": [
          "AdminTaskTemplates"
        ],
        "parameters": [
          {
            "name": "dependencyId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          },
          {
            "name": "AUGUSTUS_CORRELATION_ID",
            "in": "header"
          },
          {
            "name": "MAMRAZ_CORRELATION_ID",
            "in": "header"
          },
          {
            "name": "AUGUSTUS_CORRELATION_CREATION_TIMESTAMP",
            "in": "header"
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/api/admin/task-templates/influencers/{influencerLinkId}": {
      "delete": {
        "tags": [
          "AdminTaskTemplates"
        ],
        "parameters": [
          {
            "name": "influencerLinkId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          },
          {
            "name": "AUGUSTUS_CORRELATION_ID",
            "in": "header"
          },
          {
            "name": "MAMRAZ_CORRELATION_ID",
            "in": "header"
          },
          {
            "name": "AUGUSTUS_CORRELATION_CREATION_TIMESTAMP",
            "in": "header"
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/api/admin/task-templates/{id}/archive": {
      "post": {
        "tags": [
          "AdminTaskTemplates"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          },
          {
            "name": "AUGUSTUS_CORRELATION_ID",
            "in": "header"
          },
          {
            "name": "MAMRAZ_CORRELATION_ID",
            "in": "header"
          },
          {
            "name": "AUGUSTUS_CORRELATION_CREATION_TIMESTAMP",
            "in": "header"
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/TaskTemplate"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TaskTemplate"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/TaskTemplate"
                }
              }
            }
          }
        }
      }
    },
    "/api/admin/task-templates/{id}/unarchive": {
      "post": {
        "tags": [
          "AdminTaskTemplates"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          },
          {
            "name": "AUGUSTUS_CORRELATION_ID",
            "in": "header"
          },
          {
            "name": "MAMRAZ_CORRELATION_ID",
            "in": "header"
          },
          {
            "name": "AUGUSTUS_CORRELATION_CREATION_TIMESTAMP",
            "in": "header"
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/TaskTemplate"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TaskTemplate"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/TaskTemplate"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "ClosedListItem": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "key": {
            "maxLength": 256,
            "type": "string",
            "nullable": true
          },
          "value": {
            "maxLength": 256,
            "type": "string",
            "nullable": true
          },
          "description": {
            "maxLength": 1024,
            "type": "string",
            "nullable": true
          },
          "displayOrder": {
            "type": "integer",
            "format": "int32",
            "nullable": true
          },
          "isArchive": {
            "type": "boolean"
          },
          "creationTime": {
            "type": "string",
            "format": "date-time"
          },
          "lastUpdateTime": {
            "type": "string",
            "format": "date-time"
          }
        },
        "additionalProperties": false
      },
...
      "CreateTaskTemplateChildRequest": {
        "type": "object",
        "properties": {
          "templateKey": {
            "type": "string",
            "nullable": true
          },
          "seriesId": {
            "type": "string",
            "format": "uuid",
            "nullable": true
          },
          "categoryId": {
            "type": "string",
            "format": "uuid",
            "nullable": true
          },
          "name": {
            "type": "string",
            "nullable": true
          },
          "description": {
            "type": "string",
            "nullable": true
          },
          "notes": {
            "type": "string",
            "nullable": true
          },
          "defaultWeeksBeforeExerciseStart": {
            "type": "integer",
            "format": "int32"
          },
          "influencerIds": {
            "type": "array",
            "items": {
              "type": "string",
              "format": "uuid"
            },
            "nullable": true
          },
          "dependsOnTemplateKeys": {
            "type": "array",
            "items": {
              "type": "string"
            },
            "nullable": true
          }
        },
        "additionalProperties": false,
        "example": {
          "templateKey": "child-b",
          "name": "Child B",
          "defaultWeeksBeforeExerciseStart": 1,
          "dependsOnTemplateKeys": [
            "child-a"
          ]
        }
      },
      "CreateTaskTemplateRequest": {
        "type": "object",
        "properties": {
          "templateKey": {
            "type": "string",
            "nullable": true
          },
          "seriesId": {
            "type": "string",
            "format": "uuid",
            "nullable": true
          },
          "categoryId": {
            "type": "string",
            "format": "uuid",
            "nullable": true
          },
          "name": {
            "type": "string",
            "nullable": true
          },
          "description": {
            "type": "string",
            "nullable": true
          },
          "notes": {
            "type": "string",
            "nullable": true
          },
          "defaultWeeksBeforeExerciseStart": {
            "type": "integer",
            "format": "int32"
          },
          "influencerIds": {
            "type": "array",
            "items": {
              "type": "string",
              "format": "uuid"
            },
            "nullable": true
          },
          "dependsOnTemplateKeys": {
            "type": "array",
            "items": {
              "type": "string"
            },
            "nullable": true
          },
          "children": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/CreateTaskTemplateChildRequest"
            },
            "nullable": true
          }
        },
        "additionalProperties": false,
        "example": {
          "templateKey": "root-template",
          "seriesId": "3fa85f64-5717-4562-b3fc-2c963f66afa6",
          "categoryId": "3fa85f64-5717-4562-b3fc-2c963f66afb0",
          "name": "Root Template",
          "description": "Top level task template",
          "notes": "Optional notes",
          "defaultWeeksBeforeExerciseStart": 2,
          "influencerIds": [
            "3fa85f64-5717-4562-b3fc-2c963f66afc1"
          ],
          "children": [
            {
              "templateKey": "child-a",
              "name": "Child A",
              "defaultWeeksBeforeExerciseStart": 1
            },
            {
              "templateKey": "child-b",
              "name": "Child B",
              "defaultWeeksBeforeExerciseStart": 1,
              "dependsOnTemplateKeys": [
                "child-a"
              ]
            }
          ]
        }
      },
...
      "TaskTemplate": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "parentId": {
            "type": "string",
            "format": "uuid",
            "nullable": true
          },
          "seriesId": {
            "type": "string",
            "format": "uuid",
            "nullable": true
          },
          "categoryId": {
            "type": "string",
            "format": "uuid",
            "nullable": true
          },
          "serialNumber": {
            "type": "integer",
            "format": "int32"
          },
          "name": {
            "maxLength": 256,
            "type": "string",
            "nullable": true
          },
          "description": {
            "maxLength": 1024,
            "type": "string",
            "nullable": true
          },
          "notes": {
            "maxLength": 1024,
            "type": "string",
            "nullable": true
          },
          "defaultWeeksBeforeExerciseStart": {
            "type": "integer",
            "format": "int32"
          },
          "isArchive": {
            "type": "boolean"
          },
          "creationTime": {
            "type": "string",
            "format": "date-time"
          },
          "lastUpdateTime": {
            "type": "string",
            "format": "date-time"
          },
          "parent": {
            "$ref": "#/components/schemas/TaskTemplate"
          },
          "series": {
            "$ref": "#/components/schemas/ClosedListItem"
          },
          "category": {
            "$ref": "#/components/schemas/ClosedListItem"
          },
          "children": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/TaskTemplate"
            },
            "nullable": true
          },
          "dependencies": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/TaskTemplateDependency"
            },
            "nullable": true
          },
          "dependedOnBy": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/TaskTemplateDependency"
            },
            "nullable": true
          },
          "influencers": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/TaskTemplateInfluencer"
            },
            "nullable": true
          }
        },
        "additionalProperties": false
      },
      "TaskTemplateDependency": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "templateId": {
            "type": "string",
            "format": "uuid",
            "nullable": true
          },
          "dependsOnId": {
            "type": "string",
            "format": "uuid",
            "nullable": true
          },
          "creationTime": {
            "type": "string",
            "format": "date-time"
          },
          "template": {
            "$ref": "#/components/schemas/TaskTemplate"
          },
          "dependsOn": {
            "$ref": "#/components/schemas/TaskTemplate"
          }
        },
        "additionalProperties": false
      },
      "TaskTemplateInfluencer": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "templateId": {
            "type": "string",
            "format": "uuid",
            "nullable": true
          },
          "influencerId": {
            "type": "string",
            "format": "uuid",
            "nullable": true
          },
          "creationTime": {
            "type": "string",
            "format": "date-time"
          },
          "template": {
            "$ref": "#/components/schemas/TaskTemplate"
          },
          "influencer": {
            "$ref": "#/components/schemas/ClosedListItem"
          }
        },
        "additionalProperties": false
      },
...
      "UpdateTaskTemplateRequest": {
        "type": "object",
        "properties": {
          "seriesId": {
            "type": "string",
            "format": "uuid",
            "nullable": true
          },
          "categoryId": {
            "type": "string",
            "format": "uuid",
            "nullable": true
          },
          "name": {
            "type": "string",
            "nullable": true
          },
          "description": {
            "type": "string",
            "nullable": true
          },
          "notes": {
            "type": "string",
            "nullable": true
          },
          "defaultWeeksBeforeExerciseStart": {
            "type": "integer",
            "format": "int32"
          },
          "influencerIds": {
            "type": "array",
            "items": {
              "type": "string",
              "format": "uuid"
            },
            "nullable": true
          }
        },
        "additionalProperties": false
      },
...
    },
    "securitySchemes": {
      "SessionId": {
        "type": "apiKey",
        "name": "session-id",
        "in": "header"
      }
    }
  },
  "security": [
    {
      "SessionId": [ ]
    }
  ]
}